% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algae.R
\docType{data}
\name{algae}
\alias{algae}
\title{Algae data}
\format{An object of class \code{"data.frame"}.}
\source{
Source goes here.
}
\usage{
data(algae)
}
\description{
Algae data. More details here.
}
\examples{
data(algae)
}
\references{
References go here.
}
\keyword{datasets}
