% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biochem.R
\docType{data}
\name{biochem}
\alias{biochem}
\title{Biochem data}
\format{An object of class \code{"data.frame"}.}
\source{
Source goes here.
}
\usage{
data(biochem)
}
\description{
Biochem data. More details here.
}
\examples{
data(biochem)
}
\references{
References go here.
}
\keyword{datasets}
