% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bus.R
\docType{data}
\name{bus}
\alias{bus}
\title{Bus data}
\format{An object of class \code{"data.frame"}.}
\source{
Source goes here.
}
\usage{
data(bus)
}
\description{
Bus data. More details here.
}
\examples{
data(bus)
}
\references{
References go here.
}
\keyword{datasets}
