% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flour.R
\docType{data}
\name{flour}
\alias{flour}
\title{Flour data}
\format{An object of class \code{"data.frame"}.}
\source{
Source goes here.
}
\usage{
data(flour)
}
\description{
Flour data. More details here.
}
\examples{
data(flour)
}
\references{
References go here.
}
\keyword{datasets}
