% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass.R
\docType{data}
\name{glass}
\alias{glass}
\title{Glass data}
\format{
An object of class \code{"data.frame"}.
}
\source{
Hettich, S. and Bay, S.D. (1999), The UCI KDD Archive
http://kdd.ics.uci.edu, Irvine, CA: University of California,
Department of Information and Computer Science.
}
\usage{
data(glass)
}
\description{
Measurements of the presence of seven chemical constituents in
76 pieces of glass from nonfloat car windows.
}
\details{
Format: 76 cases and 7 continuous variables.
Description: The columns are:
1. RI refractive index
2. Na2O sodium oxide (unit measurement: weight percent in
corresponding oxide, as are the rest of attributes)
3. MgO magnesium oxide
4. Al2O3 aluminum oxide
5. SiO2 silcon oxide
6. K2O potassium oxide
7. CaO calcium oxide
}
\examples{
data(glass)
}
\keyword{datasets}
