% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiFuns.R
\name{huber}
\alias{huber}
\title{Tuning parameter the rho loss functions}
\usage{
huber(e)
}
\arguments{
\item{e}{the desired efficiency of the corresponding regression
estimator for Gaussian errors}
}
\value{
A length-1 vector with the corresponding tuning constant.
}
\description{
This function computes the tuning constant that yields an MM-regression
estimator with a desired asymptotic efficiency when computed with a
rho function in the corresponding family. The output of this
function can be passed to the functions \link{lmrobdet.control},
\link{mscale} and \link{rho}.
}
\examples{
# Tuning parameters for an 85\%-efficient M-estimator at a Gaussian model
huber(.95)

}
\author{
Kjell Konis
}
