% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.R
\docType{data}
\name{image}
\alias{image}
\title{Image data}
\format{
An object of class \code{"data.frame"}.
}
\source{
Source: Frery, A. (2005), Personal communication.
}
\usage{
data(image)
}
\description{
These data are part of a synthetic aperture satellite radar image corresponding
to a suburb of Munich, and  contain the values corresponding to three frequency
bands for each of 1573 pixels of a radar image.
}
\details{
Format: 1573 cases and 3 variables.
}
\examples{
data(image)
}
\keyword{datasets}
