% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leuk.dat.R
\docType{data}
\name{leuk.dat}
\alias{leuk.dat}
\title{Leukemia Data}
\format{An object of class \code{"data.frame"}.}
\source{
Here is the source.
}
\usage{
data(leuk.dat)
}
\description{
A description of the data
}
\details{
Description: The following features are present
wbc: wbc
ag: ag
y: y, the response.

Format: Numeric, 33 rows and 3 columns.
}
\examples{
data(leuk.dat)
}
\keyword{datasets}
