% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFPE.R
\name{lmrob.RFPE}
\alias{lmrob.RFPE}
\title{RFPE of a lmrobdet object}
\usage{
lmrob.RFPE(object, scale = NULL)
}
\arguments{
\item{object}{a robust fit as returned by \code{\link{lmrobdet}}.}

\item{scale}{an optional residual scale estimator. If \code{NULL} the residual
scale estimator that corresponds to the model in \code{object} is used.}
}
\value{
the RFPE corresponding to the fit model.
}
\description{
This function computes the RFPE for a model fit using \code{\link{lmrobdet}}.
}
\references{
\url{http://thebook}
}
\seealso{
\code{\link{lmrobdet}}
}
\author{
Victor Yohai, Matias Salibian-Barrera, \email{matias@stat.ubc.ca}
}
