% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrob.MM.R
\name{lmrob.S}
\alias{lmrob.S}
\title{Compute  S-estimator for linear model -- using  "fast S" algorithm --> ../man/lmrob.S.Rd}
\usage{
lmrob.S(x, y, control, trace.lev = control$trace.lev, mf = NULL)
}
\description{
Compute  S-estimator for linear model -- using  "fast S" algorithm --> ../man/lmrob.S.Rd
}
