% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmrob2.R
\name{lmrob2}
\alias{lmrob2}
\title{Robust DCML linear regression estimators}
\usage{
lmrob2(formula, data, subset, weights, na.action, model = TRUE,
  x = !control$compute.rd, y = FALSE, singular.ok = TRUE,
  contrasts = NULL, offset = NULL, control = lmrob2.control())
}
\arguments{
\item{formula}{a symbolic description of the model to be fit.}

\item{data}{an optional data frame, list or environment containing
the variables in the model. If not found in \code{data}, model variables
are taken from \code{environment(formula)}, which usually is the
root environment of the current R session.}

\item{subset}{an optional vector specifying a subset of observations to be used.}

\item{weights}{an optional vector of weights to be used in the fitting process.}

\item{na.action}{a function to indicates what should happen when the data contain NAs. 
The default is set by the \link{na.action} setting of \link{options}, and is 
\code{na.fail} if that is unset.}

\item{model}{logical value indicating whether to return the model frame}

\item{x}{logical value indicating whether to return the model matrix}

\item{y}{logical value indicating whether to return the vector of responses}

\item{singular.ok}{logical value. If \code{FALSE} a singular fit produces an error.}

\item{contrasts}{an optional list. See the \code{contrasts.arg} of \link{model.matrix.default}.}

\item{offset}{this can be used to specify an a priori known component to be included 
in the linear predictor during fitting. An offset term can be included in the formula 
instead or as well, and if both are specified their sum is used.}

\item{control}{a list specifying control parameters as returned by the function 
\link{lmrob2.control}.}
}
\value{
A list with the following components:
\item{coefficients}{The estimated vector of regression coefficients} 
\item{scale}{The estimated scale of the residuals}
\item{residuals}{The vector of residuals associated with the fit}
}
\description{
This function computes MM-based Distance Constrained
Maximum Likelihood regression estimators for linear models.
}
\details{
These are the details. In particular, this function computes the
DCML estimators combining the least squares estimator and a
robust MM-estimator, the latter computed using Pen~a-Yohai
candidates (instead of subsampling ones).
}
\examples{
data(coleman)
m2 <- lmrob2(Y ~ ., data=coleman)

}
\references{
\url{http://thebook}
}
\seealso{
\code{\link{lmrob}}
}
\author{
Matias Salibian-Barrera, \email{matias@stat.ubc.ca}, based on \code{lmrob}
}
