% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsRobTestMM.R
\name{lsRobTestMM}
\alias{lsRobTestMM}
\title{Test for Least Squares Bias Using Robust MM Regressions}
\usage{
lsRobTestMM(object, test = c("T", "T0"), ...)
}
\arguments{
\item{object}{An MM regression fitted model whose class is *lmrobdetMM*.}

\item{test}{A character vector indicating which of two type of tests "T" or 
"T0: are used, with type "T" the default.}

\item{...}{Pass through parameters}
}
\value{
A list with component names coefs, full, test, efficiency
}
\description{
Test for Least Squares Bias Using Robust MM Regressions
}
\details{
The original version of \code{lsRobTestMM} is the \code{lsRobTest}
in the package *robust*. The function \code{lsRobTest} had options *T1* and
*T2*. However, we only recommend *T2*, and deprecate *T1*. Accordingly we
use *T* for the former *T2*, and use *T0* for the former *T1*, and we
deprecate *T0*.

The *coefs* component of the list is a matrix whose row names are
the names of the regression predictor variables, and whose columns *LS*, 
*Robust*, *Delta*, *Std.error*, *t-stat*, *p-value* contain respectively,
the least squares and robust coefficient estimates, the differences in the
coefficient estimates, the standard errors of the differences, the resulting
t-statistic values, and the resulting z-test p-values.

The *full* component of the list is itself a list with components the full
model quadratic form chi-squared statistic value (*stat*), the degrees of 
freedom (*df*), and the full model p value (*p.value*).

The *test* component of the list is a character value indicating which of the 
tests *T* and *T0* has been computed.

The *efficiency* component of the list is *NULL* when test *T* has been used,
and is equal to the normal distribution efficiency of the *lmrobdetMM*
estimate when test *T0* has been used.
}
\examples{
args(lsRobTestMM)
}
\author{
Kjell Konis
}
