% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mineral.R
\docType{data}
\name{mineral}
\alias{mineral}
\title{Mineral data}
\format{An object of class \code{"data.frame"}.}
\source{
Source goes here.
}
\usage{
data(mineral)
}
\description{
Mineral data. More details here.
}
\examples{
data(mineral)
}
\references{
References go here.
}
\keyword{datasets}
