% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralgia.R
\docType{data}
\name{neuralgia}
\alias{neuralgia}
\title{Neuralgia data}
\format{An object of class \code{"data.frame"}.}
\source{
Source goes here.
}
\usage{
data(neuralgia)
}
\description{
Neuralgia data. More details here.
}
\examples{
data(neuralgia)
}
\references{
References go here.
}
\keyword{datasets}
