% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oats.R
\docType{data}
\name{oats}
\alias{oats}
\title{Oats data}
\format{An object of class \code{"data.frame"}.}
\source{
Source goes here.
}
\usage{
data(oats)
}
\description{
Oats data. More details here.
}
\examples{
data(oats)
}
\references{
References go here.
}
\keyword{datasets}
