% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openShinyUI.R
\name{openShinyUI}
\alias{openShinyUI}
\title{Open the Shiny interface for the package}
\usage{
openShinyUI()
}
\description{
This function opens the Shiny interface for the package.
}
\references{
\url{http://thebook}
}
\author{
Gregory Brownson \email{gsb25@uw.edu}
}
