% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiFuns.R
\name{opt}
\alias{opt}
\title{Tuning parameter for a rho function in the (asymptotic bias-) optimal family}
\usage{
opt(e)
}
\arguments{
\item{e}{the desired efficiency of the corresponding regression
estimator for Gaussian errors}
}
\value{
A vector with named elements containing the corresponding tuning
parameters.
}
\description{
This function computes the tuning constant that yields an MM-regression
estimator with a desired asymptotic efficiency when computed with a
rho function in the corresponding family. The output of this
function can be passed to the functions \link{lmrobdet.control},
\link{mscale} and \link{rho}.
}
\examples{
# Tuning parameters for an 85\%-efficient M-estimator at a Gaussian model
opt(.85)

}
\author{
Kjell Konis
}
