% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.lsRobTest.R
\name{print.lsRobTest}
\alias{print.lsRobTest}
\title{Print an lsRobTest Object}
\usage{
\method{print}{lsRobTest}(x, digits = 4, ...)
}
\arguments{
\item{x}{lmrobdetMM fitted model object}

\item{digits}{significant digits printed, default digits = 4}

\item{...}{pass through parameters}
}
\value{
print selected components of lmrobdetMM object
}
\description{
Print an lsRobTest Object
}
