% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resex.R
\docType{data}
\name{resex}
\alias{resex}
\title{Resex data}
\format{An object of class \code{"data.frame"}.}
\source{
Source goes here.
}
\usage{
data(resex)
}
\description{
Resex data. More details here.
}
\examples{
data(resex)
}
\references{
References go here.
}
\keyword{datasets}
