% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiFuns.R
\name{rho}
\alias{rho}
\title{Tukey bisquare rho function}
\usage{
rho(u, family, normalize = TRUE)
}
\arguments{
\item{u}{point or vector at which rho is to be evaluated}

\item{cc}{tuning parameter}
}
\value{
The value of \code{rho_cc} at \code{u}
}
\description{
Tukey bisquare rho function
}
\author{
Matias Salibian-Barrera, \email{matias@stat.ubc.ca}
}
