% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCML.R
\name{rhoprime}
\alias{rhoprime}
\title{The first derivative of Tukeys bisquare rho function}
\usage{
rhoprime(r, cc)
}
\arguments{
\item{r}{point or vector at which the derivative of rho is to be evaluated}

\item{cc}{tuning parameter}
}
\value{
The value of the first derivative \code{rho_cc} at \code{r}
}
\description{
The first derivative of Tukeys bisquare rho function
}
\author{
Matias Salibian-Barrera, \email{matias@stat.ubc.ca}
}
