% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiFuns.R
\name{rhoprime2}
\alias{rhoprime2}
\title{The second derivative of Tukey bisquare rho function}
\usage{
rhoprime2(u, family, cc, standardize = FALSE)
}
\arguments{
\item{u}{scalar or vector at which the second derivative of rho is to be evaluated}

\item{family}{tuning parameters as computed by 
a rho family function. Currently the following are implemented
\code{\link{bisquare}}, \code{\link{optimal}}, and \code{\link{modified.optimal}}.}
}
\value{
The value of the second derivative of \code{rho} evaluated at \code{u}
}
\description{
The second derivative of Tukey bisquare rho function
}
\author{
Matias Salibian-Barrera, \email{matias@stat.ubc.ca}
}
