% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyUI.R
\name{ShinyUI}
\alias{ShinyUI}
\title{Open the Shiny interface for the package}
\usage{
ShinyUI()
}
\description{
This function opens the Shiny interface for the package.
}
\references{
\url{http://www.wiley.com/go/maronna/robust}
}
\author{
Gregory Brownson \email{gsb25@uw.edu}
}
