% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shock.R
\docType{data}
\name{shock}
\alias{shock}
\title{Shock data}
\format{An object of class \code{"data.frame"}.}
\source{
Source goes here.
}
\usage{
data(shock)
}
\description{
Shock data. More details here.
}
\examples{
data(shock)
}
\references{
References go here.
}
\keyword{datasets}
