% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skin.R
\docType{data}
\name{skin}
\alias{skin}
\title{Skin data}
\format{An object of class \code{"data.frame"}.}
\source{
Source goes here.
}
\usage{
data(skin)
}
\description{
Skin data. More details here.
}
\examples{
data(skin)
}
\references{
References go here.
}
\keyword{datasets}
