% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackloss.R
\docType{data}
\name{stackloss}
\alias{stackloss}
\title{Stackloss data}
\format{
An object of class \code{"data.frame"}.
}
\source{
Brownlee, K.A. (1965), Statistical Theory and Methodology in Science and
Engineering, 2nd Edition, New York: John Wiley & Sons, Inc.
}
\usage{
data(stackloss)
}
\description{
Observations from 21 days operation of a plant for the oxidation
of ammonia as a stage in the production of nitric acid.
}
\details{
Format: 21 cases and 4 continuous variables.
Description: The columns are:
1. air flow
2. cooling water inlet temperature (C)
3. acid concentration (%)
4. Stack loss, defined as the percentage of ingoing ammonia
that escapes unabsorbed (response)
}
\examples{
data(stackloss)
}
\keyword{datasets}
