% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFPE.R
\name{step.lmrobdet}
\alias{step.lmrobdet}
\title{Robust stepwise using RFPE}
\usage{
step.lmrobdet(object, scope, direction = c("both", "backward", "forward"),
  trace = TRUE, keep = NULL, steps = 1000, whole.path = FALSE)
}
\arguments{
\item{object}{a robust fit as returned by \code{\link{lmrobdet}}}

\item{scope}{defines the range of models to be examined. This should be either a single formula, or a list containing components upper and lower, both formulae. See the details below.}

\item{direction}{the direction of stepwise search. Currenly only \code{backward} is implemented.}

\item{trace}{if \code{TRUE} information about each step is printed on the screen.}

\item{keep}{a filter function whose input is a fitted model object and the associated AIC statistic, and whose output is arbitrary. Typically keep will select a subset of the components of the object and return them. The default is not to keep anything.}

\item{steps}{maximum number of steps to be performed. Defaults to 1000, which should mean as many as needed.}

\item{whole.path}{if \code{FALSE} (default) variables are dropped until the RFPE fails to improve. If \code{TRUE} the best variable to be dropped is removed, even if this does not improve the RFPE.}
}
\value{
either a robust fit as obtained by \code{lmrobdet} using the final model, or a list of fits one for each step in the process.
}
\description{
This function performs stepwise variable selection using the RFPE
criterion and the robust regression estimators computed with
\code{\link{lmrobdet}}.
}
\references{
\url{http://thebook}
}
\seealso{
\code{\link{DCML}}, \code{\link{MMPY}}, \code{\link{SMPY}}
}
\author{
Victor Yohai, Matias Salibian-Barrera, \email{matias@stat.ubc.ca}
}
