% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicle.R
\docType{data}
\name{vehicle}
\alias{vehicle}
\title{Vehicle data}
\format{An object of class \code{"data.frame"}.}
\source{
Source goes here.
}
\usage{
data(vehicle)
}
\description{
Vehicle data. More details here.
}
\examples{
data(vehicle)
}
\references{
References go here.
}
\keyword{datasets}
