% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waste.R
\docType{data}
\name{waste}
\alias{waste}
\title{Waste data}
\format{
An object of class \code{"data.frame"}.
}
\source{
Golueke, C.G. and McGauhey, P.H. (1970), Comprehensive Studies of
Solid Waste Management, US Department of Health, Education and Welfare, Public
Health Services Publication No. 2039.
}
\usage{
data(waste)
}
\description{
Waste data. The original data are the result of a study on production
waste and land use by Golueke and McGauhey (1970), and contain nine variables,
of which we consider six.
}
\details{
Format: 40 cases and 6 continuous variables.
Description: The columns are
1. industrial land (acres)
2. fabricated metals (acres)
3. trucking and wholesale trade (acres)
4. retail trade (acres)
5. restaurants and hotels (acres)
6. solid waste (millions of tons), response
}
\examples{
data(waste)
}
\references{
Golueke, C.G. and McGauhey, P.H. (1970), Comprehensive Studies of
Solid Waste Management, US Department of Health, Education and Welfare, Public
Health Services Publication No. 2039.
}
\keyword{datasets}
