% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wine.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine data}
\format{An object of class \code{"data.frame"}.}
\source{
Source goes here.
}
\usage{
data(wine)
}
\description{
Wine data. More details here.
}
\examples{
data(wine)
}
\references{
References go here.
}
\keyword{datasets}
