% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biochem.R
\docType{data}
\name{biochem}
\alias{biochem}
\title{Biochem data}
\format{An object of class \code{"data.frame"}.}
\source{
Seber, G.A.F. (1984), Multivariate Observations. New York: John Wiley.
}
\usage{
data(biochem)
}
\description{
Two biochemical measurements on 12 men with similar weights.
}
\details{
Format: Numeric, 12 rows, two columns
}
\examples{
data(biochem)
}
\keyword{datasets}
