% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flour.R
\docType{data}
\name{flour}
\alias{flour}
\title{Flour data}
\format{An object of class \code{"data.frame"}.}
\source{
Analytical Methods Committee (1989), Robust statistics-How not to reject
outliers, Analyst, 114, 1693-1702.
}
\usage{
data(flour)
}
\description{
Determinations of the copper content in wholemeal flour
(in parts per million), sorted in ascending order.
Format: numeric vector of size 24.
}
\examples{
data(flour)
}
\references{
References go here.
}
\keyword{datasets}
