% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INVTR2.R
\name{INVTR2}
\alias{INVTR2}
\title{Robust R^2 coefficient of determination}
\usage{
INVTR2(RR2, cc)
}
\arguments{
\item{RR2}{the proportional difference in loss functions (a naive robust R^2 coefficient).}

\item{cc}{the tuning constant for the rho function}
}
\value{
An unbiased version of the robust R^2 coefficient of determination.
}
\description{
This function computes a robust version of the R^2 coefficient of determination.
}
\details{
This function computes a robust version of the R^2 coefficient. It currently only
works for MM regression estimators computed with a rho function in Tukey's 
bisquare family.
}
\references{
\url{http://thebook}
}
\author{
Victor Yohai
}
