% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFPE.R
\name{lmrobdet.RFPE}
\alias{lmrobdet.RFPE}
\title{RFPE of a model fit with \code{\link{lmrobdet}}}
\usage{
lmrobdet.RFPE(object, scale = NULL)
}
\arguments{
\item{object}{the \code{MM} element (of class \code{\link{lmrob}}) in an object of class \code{\link{lmrobdet}}.}

\item{scale}{an optional residual scale estimator. If \code{NULL} the residual
scale estimator that corresponds to the model in \code{object} is used.}
}
\value{
the RFPE corresponding to the fit model.
}
\description{
This function computes the RFPE for the MM-estimator obtained using \code{\link{lmrobdet}}.
}
\references{
\url{http://thebook}
}
\seealso{
\code{\link{lmrobdet}}
}
\author{
Victor Yohai, Matias Salibian-Barrera, \email{matias@stat.ubc.ca}
}
