% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsRobTestMM.R
\name{lsRobTestMM}
\alias{lsRobTestMM}
\title{Test for Differences between LS and Robust MM Regressions}
\usage{
lsRobTestMM(object, test = c("T2", "T1"), ...)
}
\arguments{
\item{object}{An MM regression fitted model whose class is "lmrobdetMM".}

\item{test}{A character vector indicating which of two type of tests "T1" or 
"T2: are used, with type "T2" the default.}

\item{...}{Pass through parameters}
}
\value{
A list with component names coefs, full, test, efficiency
}
\description{
Test for Differences between LS and Robust MM Regressions
}
\details{
The *coefs* component of the list is a matrix whose row names are
the names of the regression predictor variables, and whose columns *LS*, 
*Robust*, *Delta*, *Std.error*, *t-stat*, *p-value* contain respectively,
the least squares and robust coefficient estimates, the differences in the
coefficient estimates, the standard errors of the differences, the resulting
t-statistic values, and the resulting z-test p-values.

The *full* component of the list is itself a list with components the full
model quadratic form chi-squared statistic value (*stat*), the degrees of 
freedom (*df*), and the full model p value (*p.value*).

The test component . . . 

The efficiency component . . .
}
\examples{
args(lsRobTestMM)
}
