% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLocDis.R
\name{locScaleM}
\alias{locScaleM}
\alias{MLocDis}
\title{Robust univariate location and scale M-estimators}
\usage{
locScaleM(x, psi = "Bis", eff = 0.9, maxit = 50, tol = 1e-04)
}
\arguments{
\item{x}{a vector of univariate observations}

\item{psi}{a string indicating which score function to use. Valid options are "Bis" for
bi-square and "Hub" for a Huber-type.}

\item{eff}{desired asymptotic efficiency. Valid options are 0.9 (default), 0.85 and 0.95.}

\item{maxit}{maximum number of iterations allowed.}

\item{tol}{tolerance to decide convergence of the iterative algorithm.}
}
\value{
A list with the following components:
\item{mu}{The location estimator}
\item{std.mu}{Estimated standard deviation of the location estimator \code{mu}}
\item{disper}{M-scale/dispersion estimator}
}
\description{
This function computes M-estimators for location and scale.
}
\details{
This function computes M-estimators for location and scale.
}
\references{
\url{http://thebook}
}
\author{
Ricardo Maronna, \email{rmaronna@retina.ar}
}
