% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiRobu.R
\name{MMultiSHR}
\alias{MMultiSHR}
\title{MMultiSHR robust multivariate location and scatter estimator}
\usage{
MMultiSHR(X, maxit = 50, tolpar = 1e-04)
}
\arguments{
\item{X}{a data matrix with observations in rows.}

\item{maxit}{Maximum number of iterations.}

\item{tolpar}{Tolerance to decide converngence.#'}
}
\value{
A list with the following components:
\item{mu}{The location estimator}
\item{V}{The scatter matrix estimator, scaled for consistency at the normal distribution}
\item{dista}{Robust Mahalanobis distances}
}
\description{
This function computes the MMultiSHR robust estimator for multivariate location and scatter.
}
\details{
This function computes the MMultiSHRrobust estimator for multivariate location and scatter.
}
\references{
\url{http://thebook}
}
\author{
Ricardo Maronna, \email{rmaronna@retina.ar}
}
