% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiRobu.R
\name{covRobMM}
\alias{covRobMM}
\alias{MMultiSHR}
\title{MM robust multivariate location and scatter estimator}
\usage{
covRobMM(X, maxit = 50, tolpar = 1e-04)
}
\arguments{
\item{X}{a data matrix with observations in rows.}

\item{maxit}{Maximum number of iterations.}

\item{tolpar}{Tolerance to decide converngence.#'}
}
\value{
A list with the following components:
\item{mu}{The location estimator}
\item{V}{The scatter matrix estimator, scaled for consistency at the normal distribution}
\item{center}{The location estimator. Same as \code{mu} above.}
\item{cov}{The scatter matrix estimator, scaled for consistency at the normal distribution. Same as \code{V} above.}
\item{dista}{Robust Mahalanobis distances}
}
\description{
This function computes the MM robust estimator for multivariate location and scatter with the "SHR" loss function.
}
\details{
This function computes the MM robust estimator for multivariate location and scatter with the "SHR" loss function.
}
\references{
\url{http://thebook}
}
\author{
Ricardo Maronna, \email{rmaronna@retina.ar}
}
