% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oats.R
\docType{data}
\name{oats}
\alias{oats}
\title{Oats data}
\format{
An object of class \code{"data.frame"}.
}
\source{
Scheffe, H. (1959), Analysis of Variance. New York: John Wiley.
}
\usage{
data(oats)
}
\description{
Yield of grain for eight varieties of oats in five replications of a randomized-block experiment
}
\details{
Format: Two-way ANOVA table with 8 rows and 5 columns.
}
\examples{
data(oats)
}
\references{
References go here.
}
\keyword{datasets}
