% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resex.R
\docType{data}
\name{resex}
\alias{resex}
\title{Resex data}
\format{An object of class \code{"data.frame"}.}
\source{
Source Engineering, 2nd. Edition, New York, John Wiley.
}
\usage{
data(resex)
}
\description{
A monthly series of inward movement of residential telephone extensions in a
fixed geographic area from January 1966 to May 1973.
}
\details{
Format: numeric vector of size 89.
}
\examples{
data(resex)
}
\references{
Brubacher. S.R. (1974), Time series outlier detection and modeling
with interpolation, Bell Laboratories Technical Memo.
}
\keyword{datasets}
