% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiFuns.R
\name{bisquare}
\alias{bisquare}
\alias{optimal}
\alias{modified.optimal}
\alias{rho}
\title{Tukey bisquare rho object}
\usage{
bisquare(efficiency, breakdown.point)

optimal(e)

modified.optimal(e)

rho(u, family, cc, standardize = TRUE)
}
\arguments{
\item{efficiency}{the desired efficiency of the corresponding regression 
estimator for Gaussian errors}

\item{breakdown.point}{the desired breakdown of the associated regression 
estimator}

\item{e}{the desired efficiency of the corresponding regression 
estimator for Gaussian errors}

\item{u}{point or vector at which rho is to be evaluated}

\item{family}{tuning parameters as computed by 
a rho family function. Currently the following are implemented
\code{\link{bisquare}}, \code{\link{optimal}}, and \code{\link{modified.optimal}}.}

\item{e}{the desired efficiency of the corresponding regression 
estimator for Gaussian errors}
}
\value{
A list with elements \code{name} (the string 'bisquare') and the corresponding tuning
constant \code{cc}

A list with elements \code{name} (the string 'optimal') and the corresponding tuning
constant \code{cc}

A list with elements \code{name} (the string 'modified.optimal') and the corresponding tuning
constant \code{cc}

The value of \code{rho} at \code{u}
}
\description{
Tukey bisquare rho object

Optimal rho function object

Modified optimal rho function object

Tukey bisquare rho function
}
\author{
Kjell Konis

Kjell Konis

Kjell Konis

Matias Salibian-Barrera, \email{matias@stat.ubc.ca}
}
