% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiFuns.R
\name{bisquare}
\alias{bisquare}
\alias{optimal}
\alias{modified.optimal}
\alias{rho}
\title{Tukey bisquare rho object}
\usage{
bisquare(efficiency)

optimal(e)

modified.optimal(e)

rho(u, family = " bisquare", cc, standardize = TRUE)
}
\arguments{
\item{efficiency}{the desired efficiency of the corresponding regression
estimator for Gaussian errors}

\item{e}{the desired efficiency of the corresponding regression
estimator for Gaussian errors}

\item{u}{point or vector at which rho is to be evaluated}

\item{family}{family string specifying the name of the family of loss function to be used (current valid
options are "bisquare", "optimal" and "modified.optimal").}

\item{cc}{tuning parameters to be computed according to efficiency and / or breakdown 
considerations. See \link{lmrobdet.control}.}

\item{standardize}{logical value determining whether the rho function is to be
standardized so that its maximum value is 1. See \link{Mpsi}.}

\item{breakdown.point}{the desired breakdown of the associated regression
estimator}

\item{e}{the desired efficiency of the corresponding regression
estimator for Gaussian errors}
}
\value{
A list with elements \code{name} (the string 'bisquare') and the corresponding tuning
constant \code{cc}

A list with elements \code{name} (the string 'optimal') and the corresponding tuning
constant \code{cc}

A list with elements \code{name} (the string 'modified.optimal') and the corresponding tuning
constant \code{cc}

The value(s) of \code{rho} at \code{u}
}
\description{
Tukey bisquare rho object

Optimal rho function object

Modified optimal rho function object

Tukey bisquare rho function
}
\author{
Kjell Konis

Kjell Konis

Kjell Konis

Matias Salibian-Barrera, \email{matias@stat.ubc.ca}
}
