% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psiFuns.R
\name{rhoprime}
\alias{rhoprime}
\title{The first derivative of Tukeys bisquare rho function}
\usage{
rhoprime(u, family, cc, standardize = FALSE)
}
\arguments{
\item{u}{scalar or vector at which the derivative of rho is to be evaluated}

\item{family}{tuning parameters as computed by 
a rho family function. Currently the following are implemented
\code{\link{bisquare}}, \code{\link{optimal}}, and \code{\link{modified.optimal}}.}
}
\value{
The value of the first derivative \code{rho} evaluated at \code{u}
}
\description{
The first derivative of Tukeys bisquare rho function
}
\author{
Matias Salibian-Barrera, \email{matias@stat.ubc.ca}
}
