% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCML.R, R/psiFuns.R
\name{MMPY}
\alias{MMPY}
\alias{rhoprime}
\title{The first derivative of Tukeys bisquare rho function}
\usage{
MMPY(X, y, control, mf)

rhoprime(u, family, standardize = FALSE)
}
\arguments{
\item{X}{design matrix}

\item{y}{response vector}

\item{control}{a list of control parameters as returned by \code{\link{lmrobdet.control}}}

\item{mf}{model frame}

\item{u}{scalar or vector at which the derivative of rho is to be evaluated}

\item{cc}{tuning parameter}

\item{r}{scalar or vector at which the derivative of rho is to be evaluated}

\item{cc}{tuning parameter}
}
\value{
The value of the first derivative \code{rho_cc} evaluated at \code{r}

an \code{\link{lmrob}} object witht the M-estimator
obtained starting from the S-estimator computed with the
Pen~a-Yohai initial candidates. The properties of the final
estimator (efficiency, etc.) are determined by the tuning constants in
the argument \code{control}.

The value of the first derivative \code{rho_cc} evaluated at \code{r}
}
\description{
The first derivative of Tukeys bisquare rho function

The first derivative of Tukeys bisquare rho function
}
\references{
\url{http://thebook}
}
\seealso{
\code{\link{DCML}}, \code{\link{MMPY}}, \code{\link{SMPY}}
}
\author{
Matias Salibian-Barrera, \email{matias@stat.ubc.ca}

Victor Yohai, Matias Salibian-Barrera, \email{matias@stat.ubc.ca}

Matias Salibian-Barrera, \email{matias@stat.ubc.ca}
}
