% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCML.R
\name{rhoprime2}
\alias{rhoprime2}
\title{The second derivative of Tukey bisquare rho function}
\usage{
rhoprime2(r, cc)
}
\arguments{
\item{r}{point or vector at which the second derivative of rho is to be evaluated}

\item{cc}{tuning parameter}
}
\value{
The value of the second derivative of \code{rho_cc} at \code{r}
}
\description{
The second derivative of Tukey bisquare rho function
}
\author{
Matias Salibian-Barrera, \email{matias@stat.ubc.ca}
}
