% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiRobu.R
\name{RockeMulti}
\alias{RockeMulti}
\title{Rocke's robust multivariate location and scatter estimator}
\usage{
RockeMulti(X, initial = "K", maxsteps = 5, propmin = 2, qs = 2,
  maxit = 50, tol = 1e-04)
}
\arguments{
\item{X}{a data matrix with observations in rows.}

\item{initial}{A character indicating the initial estimator. Valid options are 'K' (default) 
for a fast and deterministic one, and 'mve' for the Minimum Volume Ellipsoid.}

\item{maxsteps}{Maximum number of steps for the line search section of the algorithm.}

\item{propmin}{Proportion of weights computed from the initial estimator that will be different 
from zero. The number of observations with non-zero weights will be at least p (the number of
columns of \code{X} times \code{propmin}.}

\item{qs}{Tuning paramater for Rocke's loss functions.}

\item{maxit}{Maximum number of iterations.}

\item{tol}{Tolerance to decide converngence.#'}
}
\value{
A list with the following components:
\item{mu}{The location estimator}
\item{V}{The scatter matrix estimator, scaled for consistency at the normal distribution}
\item{dista}{Robust Mahalanobis distances}
\item{w}{weights}
\item{gamma}{gamma}
}
\description{
This function computes Rocke's robust estimator for multivariate location and scatter.
}
\details{
This function computes Rocke's robust estimator for multivariate location and scatter.
}
\references{
\url{http://thebook}
}
\author{
Ricardo Maronna, \email{rmaronna@retina.ar}
}
