% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shock.R
\docType{data}
\name{shock}
\alias{shock}
\title{Shock data}
\format{
An object of class \code{"data.frame"}.
}
\source{
Bond, N.W. (1979), Impairment of shuttlebox avoidance-learning following repeated
alcohol withdrawal episodes in rats, Pharmacology, Biochemistry and Behavior,
11, 589-591.
}
\usage{
data(shock)
}
\description{
Times recorded for a rat to go through a shuttlebox in successive attempts. If the
time exceeded 5 seconds, the rat received an electric shock for the duration of the
next attempt. The data are the number of shocks received and the average time for
all attempts between shocks.
}
\details{
Format: Numeric matrix with 16 rows and 2 columns
}
\examples{
data(shock)
}
\references{
References go here.
}
\keyword{datasets}
