% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wine.R
\docType{data}
\name{wine}
\alias{wine}
\title{Wine data}
\format{An object of class \code{"data.frame"}.}
\source{
Hettich, S. and Bay, S.D. (1999), The UCI KDD Archive http://kdd.ics.uci.edu.
Irvine, CA: University of California, Department of Information and Computer Science.
}
\usage{
data(wine)
}
\description{
It contains, for each of 59 wines grown in the same region in Italy, the quantities of 13
constituents. The original purpose of the analysis was to classify wines from different
cultivars by means of these measurements. In this example we treat cultivar one.
}
\details{
Format: Numeric, 59 rows and 13 columns. Description: The attributes are:
1. Alcohol
2. Malic acid
3. Ash
4. Alcalinity of ash
5. Magnesium
6. Total phenols
7. Flavanoids
8. Nonflavanoid phenols
9. Proanthocyanins
10. Color intensity
11. Hue
12. OD280/OD315 of diluted wines
13. Proline
}
\examples{
data(wine)
}
\keyword{datasets}
